refactor(policy): split tiers compiler facade
