+++ crates/assay-policy/src/tiers_next/mod.rs
@@ -1,0 +1,13 @@
+mod classifier;
+mod compiler;
+mod maps;
+mod types;
+
+pub use compiler::compile;
+pub use types::{
+    CidrRule, CompilationStats, CompiledPolicy, DestRule, FilePolicy, GlobRule, InodeRule,
+    NetworkPolicy, PathRule, Policy, PortRule, ProcessPolicy, Tier1Rules, Tier2Rules,
+};
+
+#[cfg(test)]
+mod tests;
